refactor: add routing & separate out MemberList component
- Move SearchBar from Navbar to MemberList
