+++ 6_useeffect_and_api/src/components/Navbar.js
@@ -1,15 +1,12 @@
 import logo from "../assets/logo.jpeg";
 
-import SearchBar from "./SearchBar";
-
-const Navbar = ({ APIData, setFilteredMembers, setNoResults }) => {
+const Navbar = () => {
 	return (
 		<div className="navbar">
 			<div className="logo">
 				<img src={logo} alt="logo"></img>
 				<h1 className="title">Web Pirates</h1>
 			</div>
-			<SearchBar APIData={APIData} setFilteredMembers={setFilteredMembers} setNoResults={setNoResults} />
 		</div>
 	)
 };
