+++ 6_useeffect_and_api/src/style.css
@@ -40,7 +40,10 @@ a {
 }
 
 .search-container {
-	width: 350px;
+	margin-top: 40px;
+	margin-bottom: 10px;
+	display: flex;
+	justify-content: center;
 }
 
 .search-bar {
@@ -51,7 +54,7 @@ a {
 	font-size: 14px;
 	height: 40px;
 	line-height: 24px;
-	width: 100%;
+	width: 350px;
 }
 
 /* content - card */
